[KOGITO-9147] Make Java service easier to override
There are two new class. Users might extend from ServiceWorkItemHandler
to implement the complete invocation mechanism or extend from
StatisServiceworkItemHandler and override the way the service instance
is retrieved.
